[server] Seperate learn contract related code
